% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_report_template.R
\name{copy_report_template}
\alias{copy_report_template}
\title{Copy Quarto Report Template}
\usage{
copy_report_template(template_name = "report_template.qmd", output_dir = ".")
}
\arguments{
\item{template_name}{Name of the template file (default:
"report_template.qmd").}
}
\description{
This function copies the Quarto (.qmd) report template from the package to
the current directory.
}
