% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_spatial_clusters.R
\name{create_spatial_clusters}
\alias{create_spatial_clusters}
\title{Create spatial clusters}
\usage{
create_spatial_clusters(
  spe,
  image_col = "Image",
  phenotype_col = "HierarchyLevel4",
  centroid_x_col = "Cell.X.Position",
  centroid_y_col = "Cell.Y.Position",
  k_closest = 10,
  max_dist = 50,
  number_clusters_kmeans = 10,
  seed = 123
)
}
\arguments{
\item{image_col}{Column name in cell_df that contains image names}

\item{phenotype_col}{Column name in cell_df that contains cell type
information}

\item{centroid_x_col}{Column name in cell_df that contains x-coordinates}

\item{centroid_y_col}{Column name in cell_df that contains y-coordinates}

\item{k_closest}{The number of nearest neighbours to consider}

\item{max_dist}{The maximum distance to consider a neighbour}

\item{number_clusters_kmeans}{The number of clusters to create with kmeans}

\item{seed}{The random seed to use}

\item{cell_df}{A data frame containing cell data}

\item{keep_proportions}{Whether to keep the proportions of nearest neighbours}
}
\value{
A data frame with spatial clusters
}
\description{
Create spatial clusters based on the proportions of nearest
neighbours of cell types
}
