% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_spe_by_metadata.R
\name{filter_spe_by_metadata}
\alias{filter_spe_by_metadata}
\title{Filter a SingleCellExperiment object by metadata}
\usage{
filter_spe_by_metadata(spe, metadata_col, metadata_values, invert = FALSE)
}
\arguments{
\item{spe}{SingleCellExperiment object}

\item{metadata_col}{Column name in colData of spe to filter by}

\item{metadata_values}{List of the metadata column values to perform
filtering on}

\item{invert}{Logical indicating whether to invert the filter (default:
FALSE)}
}
\value{
Filtered SingleCellExperiment object
}
\description{
Filter a SingleCellExperiment object by metadata
}
