% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_hierarchies}
\alias{load_hierarchies}
\title{Load cell hierarchies from YAML file}
\usage{
load_hierarchies(hierarchy_file)
}
\arguments{
\item{hierarchy_file}{Path to the YAML file containing cell hierarchies}
}
\value{
A data frame with the hierarchy levels as columns
}
\description{
Load cell hierarchies from a YAML file and return a data frame
with the hierarchy levels as columns.
}
