% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_spe_from_expr_data.R
\name{make_spe_from_expr_data}
\alias{make_spe_from_expr_data}
\title{Create a SpatialExperiment object from expression data}
\usage{
make_spe_from_expr_data(
  expression_file,
  hierarchy_file,
  metadata_cols = c("Image"),
  marker_col = "Class",
  centroid_x_col = "Centroid X",
  centroid_y_col = "Centroid Y",
  are_markers_split = FALSE,
  sample_id_col = NULL
)
}
\arguments{
\item{expression_file}{Path to the expression data file}

\item{hierarchy_file}{Path to the YAML file containing cell hierarchies}

\item{metadata_cols}{Character vector of metadata columns (default:
c("Image"))}

\item{marker_col}{Column name for the marker class (default: "Class")}

\item{centroid_x_col}{Column name for the x-coordinate of the cell centroid
(default: "Centroid X")}

\item{centroid_y_col}{Column name for the y-coordinate of the cell centroid
(default: "Centroid Y")}

\item{are_markers_split}{Logical indicating if the marker column is already
split into multiple columns (default: FALSE)}

\item{sample_id_col}{Column name for the sample ID (default: NULL)}
}
\value{
A SpatialExperiment object
}
\description{
Takes an expression file and a data frame containing the
hierarchy levels and constructs a SpatialExperiment object.
}
