% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cells_spatially.R
\name{plot_cells_spatially}
\alias{plot_cells_spatially}
\title{Plot cell types on spatial coordinates}
\usage{
plot_cells_spatially(
  spe,
  cell_type_colname = "HierarchyLevel4",
  parent_types = NULL,
  parent_type_colname = "HierarchyLevel1",
  centroid_x_col = "Cell.X.Position",
  centroid_y_col = "Cell.Y.Position",
  pointsize = 0.5,
  colour_by = "cell_type"
)
}
\arguments{
\item{spe}{SingleCellExperiment object}

\item{cell_type_colname}{Column name in colData of spe that contains cell
types (default: "HierarchyLevel4")}

\item{parent_types}{Character vector of parent cell types to include
(default: all parent types)}

\item{parent_type_colname}{Column name in colData of spe that contains parent
cell types (default: "HierarchyLevel1")}

\item{centroid_x_col}{Column name in spatial coordinates that contains x
coordinates (default: "Centroid X")}

\item{centroid_y_col}{Column name in spatial coordinates that contains y
coordinates (default: "Centroid Y")}

\item{pointsize}{Size of points in the plot (default: 0.5)}

\item{colour_by}{Column name in colData of spe to colour points by (options:
"cell_type" (default), "cluster" or "tumour")}
}
\value{
ggplot2 object
}
\description{
Plot cell types on spatial coordinates
}
