% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cells.R
\name{plot_celltypes}
\alias{plot_celltypes}
\title{Plot cell types on spatial coordinates}
\usage{
plot_celltypes(
  spe,
  celltype_colname = "HierarchyLevel4",
  centroid_x_col = "Centroid X",
  centroid_y_col = "Centroid Y",
  pointsize = 0.5
)
}
\arguments{
\item{spe}{SingleCellExperiment object}

\item{celltype_colname}{Column name in colData of spe that contains cell
types (default: "HierarchyLevel4")}

\item{centroid_x_col}{Column name in spatial coordinates that contains x
coordinates (default: "Centroid X")}

\item{centroid_y_col}{Column name in spatial coordinates that contains y
coordinates (default: "Centroid Y")}

\item{pointsize}{Size of points in the plot (default: 0.5)}
}
\value{
ggplot2 object
}
\description{
Plot cell types on spatial coordinates
}
