% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_props_by_meta.R
\name{plot_cluster_props_by_meta}
\alias{plot_cluster_props_by_meta}
\title{Plot sample proportions by some meta-value (e.g., sample_id)}
\usage{
plot_cluster_props_by_meta(spe, meta_col = "sample_id")
}
\arguments{
\item{spe}{SingleCellExperiment object containing cluster information}

\item{meta}{Column name in colData containing meta information (default:
"sample_id")}
}
\value{
ggplot object
}
\description{
Plot the proportions of cell types in each cluster
}
