% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_compartment_measurements.R
\name{plot_compartment_measurements}
\alias{plot_compartment_measurements}
\title{Plot compartment measurements}
\usage{
plot_compartment_measurements(
  measurement_data,
  compartments = c("Cell", "Nucleus"),
  measurements = c("Area", "Length", "Circularity", "Solidity", "Max diameter",
    "Min diameter")
)
}
\arguments{
\item{measurement_data}{A data.table containing segmentation measurements}

\item{compartments}{Character vector of compartments to plot (default:
c("Cell", "Nucleus"))}

\item{measurements}{Character vector of measurements to plot for each
compartment (default: c("Area", "Length", "Circularity", "Solidity",
"Max diameter", "Min diameter"))}
}
\value{
A ggplot object containing the violin plots for each compartment
}
\description{
This function takes segmentation measurements and creates violin
plots for each compartment.
}
