% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geometries.R
\name{plot_geometries}
\alias{plot_geometries}
\title{Plot cell and nucleus geometries from segmentation GeoJSON file}
\usage{
plot_geometries(
  geojson_file,
  area_width = 500,
  area_height = 500,
  min_objects = 5
)
}
\arguments{
\item{geojson_file}{Path to the GeoJSON file containing segmentation data}

\item{area_width}{Width of the area to plot (default: 500)}

\item{area_height}{Height of the area to plot (default: 500)}

\item{min_objects}{Minimum number of cell objects required in the area
(default: 5)}
}
\value{
A ggplot object containing the cell and nucleus geometries
}
\description{
This function takes cell and nucleus geometries from a
segmentation GeoJSON file and creates a ggplot showing the outlines of the
cells and nuclei within a randomly selected area containing at least a
specified minimum number of cells.
}
