% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intensity_measurements.R
\name{plot_intensity_measurements}
\alias{plot_intensity_measurements}
\title{Plot intensity measurements for all channels from cell segmentation
GeoJSON file}
\usage{
plot_intensity_measurements(
  measurement_data,
  compartments = c("Nucleus", "Cell", "Cytoplasm", "Membrane"),
  calculation = "Median"
)
}
\arguments{
\item{calculation}{Calculation to use for intensity measurements (default:
"Median"). Options are "Mean", "Median", "Min", "Max" and "Std.Dev".}

\item{geojson_file}{Path to the GeoJSON file containing segmentation data}

\item{compartment}{Compartments to plot intensity measurements for (default:
c("Nucleus", "Cell", "Cytoplasm", "Membrane")). Must be at least one of the
following: "Cell", "Nucleus", "Cytoplasm", "Membrane".}
}
\description{
Take segmentation measurements and plot intensity measurements
for the specified compartments.
}
