% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phenotyping_qc.R
\name{plot_marker_heatmap}
\alias{plot_marker_heatmap}
\title{Plot marker heatmap}
\usage{
plot_marker_heatmap(
  spe,
  markers,
  celltypes,
  parenttypes,
  celltype_colname = "HierarchyLevel4",
  parent_colname = "HierarchyLevel2"
)
}
\arguments{
\item{spe}{SingleCellExperiment object containing marker intensities}

\item{markers}{Character vector of markers to plot and the order in which
they should be plotted}

\item{celltypes}{Character vector of cell types to plot and the order in
which they should be plotted}

\item{parenttypes}{Character vector of parent cell types to plot and the
order in which they should be plotted}

\item{celltype_colname}{Column name in colData containing cell type
information (default: "HierarchyLevel4")}

\item{parent_colname}{Column name in colData containing parent cell type
information (default: "HierarchyLevel2")}
}
\value{
A ggplot2 object
}
\description{
Plot a heatmap of mean marker intensities in cell types.
}
