% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap.R
\name{plot_umap}
\alias{plot_umap}
\title{Plot UMAP}
\usage{
plot_umap(
  spe,
  markers = NULL,
  cell_types = NULL,
  cell_type_colname = "HierarchyLevel4",
  parent_types = NULL,
  parent_colname = "HierarchyLevel1"
)
}
\arguments{
\item{spe}{SingleCellExperiment object containing marker intensities}

\item{markers}{Character vector of markers to plot and the order in which
they should be plotted}

\item{cell_types}{Character vector of cell types to plot and the order in
which they should be plotted}

\item{cell_type_colname}{Column name in colData containing cell type
information (default: "HierarchyLevel4")}

\item{parent_types}{Character vector of parent cell types to include
(default: all parent types)}

\item{parent_colname}{Column name in colData containing parent cell type
information (default: "HierarchyLevel1")}
}
\value{
A ggplot2 object
}
\description{
Plot a UMAP of cell types
}
