% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_cells}
\alias{simulate_cells}
\title{Simulate cells}
\usage{
simulate_cells(
  markers = example_markers,
  n_cells = 5000,
  width = 2000,
  height = 2000,
  seed = 123
)
}
\arguments{
\item{markers}{A character vector of marker names}

\item{n_cells}{The number of cells to simulate}

\item{width}{The width of the image}

\item{height}{The height of the image}

\item{seed}{The random seed to use}
}
\value{
A data frame with simulated cell data
}
\description{
Simulate cells for testing imcvis functions
}
