% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segmentation_geometry.R
\name{get_segmentation_geometry}
\alias{get_segmentation_geometry}
\title{Get segmentation geometry from GeoJSON file}
\usage{
get_segmentation_geometry(geojson_file, only_keep_cells = TRUE)
}
\arguments{
\item{geojson_file}{Path to the GeoJSON file containing segmentation data}

\item{only_keep_cells}{Logical indicating whether to filter the data to
only keep cells (default: TRUE). If TRUE, only features with objectType
"cell" will be kept.}
}
\value{
A list with two elements: `cell` and `nucleus`, each containing a
list of geometries.
}
\description{
This function reads a GeoJSON file containing cell segmentation
geometry and returns a list with the coordinates of the cell and nucleus
geometries. If `only_keep_cells` is TRUE, it will filter the data to only
keep features with objectType "cell". It also filters out multipolygon
geometries, as they are not supported in the current plotting function.
}
