% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_segmentation_measurements.R
\name{get_segmentation_measurements}
\alias{get_segmentation_measurements}
\title{Get segmentation data from GeoJSON file}
\usage{
get_segmentation_measurements(geojson_file, only_keep_cells = TRUE)
}
\arguments{
\item{geojson_file}{Path to the GeoJSON file containing segmentation data}

\item{only_keep_cells}{Logical indicating whether to filter the data to
only keep cells (default: TRUE). If TRUE, only features with objectType
"cell" will be kept.}
}
\value{
A data.table object
}
\description{
This function reads a GeoJSON file containing cell segmentation
measurements and returns a data.table with the properties of the features.
If no measurements are present, it will return a data.table with only the
object types.
}
