% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{make_spe_from_expr_data}
\alias{make_spe_from_expr_data}
\title{Create a SpatialExperiment object from expression data}
\usage{
make_spe_from_expr_data(expression_file, hierarchy_df)
}
\arguments{
\item{expression_file}{Path to the expression data file}

\item{hierarchy_df}{Data frame containing the hierarchy levels}
}
\value{
A SpatialExperiment object
}
\description{
Create a SpatialExperiment object from expression data and
metadata.
}
