% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cluster_cell_props.R
\name{plot_cluster_cell_props}
\alias{plot_cluster_cell_props}
\title{Plot cell type proportions in each cluster}
\usage{
plot_cluster_cell_props(
  spe,
  cell_types = NULL,
  cell_type_colname = "HierarchyLevel4",
  plot_type = "bar"
)
}
\arguments{
\item{spe}{SingleCellExperiment object containing cluster information}

\item{cell_types}{Character vector of cell types to plot and the order in
which they should be plotted (default: plot all cell types)}

\item{cell_type_colname}{Column name in colData containing cell type
information (default: "HierarchyLevel4")}

\item{plot_type}{Type of plot to generate ("bar" (default) or "heatmap")}
}
\value{
ggplot object
}
\description{
Plot the proportions of cell types in each cluster
}
