% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_geometries.R
\name{plot_geometries}
\alias{plot_geometries}
\title{Plot cell and nucleus geometries from segmentation GeoJSON file}
\usage{
plot_geometries(
  geom_data,
  area_width = 500,
  area_height = 500,
  min_objects = 10,
  image = NULL
)
}
\arguments{
\item{geom_data}{A list containing cell and nucleus geometries, as
returned by `get_segmentation_geometry()`}

\item{area_width}{Width of the area to plot (default: 500)}

\item{area_height}{Height of the area to plot (default: 500)}

\item{min_objects}{Minimum number of cell objects required in the area
(default: 10)}

\item{image}{Optional raster image to use as background (default: NULL)}
}
\value{
A ggplot object containing the cell and nucleus geometries
}
\description{
This function takes cell and nucleus geometries from a list and
creates a ggplot showing the outlines of the cells and nuclei within a
randomly selected area containing at least a specified minimum number of
cells.
}
