% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_intensity_measurements.R
\name{plot_intensity_measurements}
\alias{plot_intensity_measurements}
\title{Plot intensity measurements for all channels from cell segmentation
GeoJSON file}
\usage{
plot_intensity_measurements(
  geojson_file,
  compartment = "Cell",
  calculation = "Median"
)
}
\arguments{
\item{geojson_file}{Path to the GeoJSON file containing segmentation data}

\item{compartment}{Compartment to plot intensity measurements for (default:
"Cell"). Options are "Cell", "Nucleus", "Cytoplasm", and "Membrane".}

\item{calculation}{Calculation to use for intensity measurements (default:
"Median"). Options are "Mean", "Median", "Min", "Max" and "Std.Dev".}
}
\description{
Take a GeoJSON file containing cell segmentation output and
plot intensity measurements for the specified compartment.
}
