% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_marker_heatmap.R
\name{plot_marker_heatmap}
\alias{plot_marker_heatmap}
\title{Plot marker heatmap}
\usage{
plot_marker_heatmap(
  spe,
  markers = NULL,
  celltypes = NULL,
  parenttypes = NULL,
  celltype_colname = "HierarchyLevel4",
  parent_colname = "HierarchyLevel2",
  value = "expression"
)
}
\arguments{
\item{spe}{SingleCellExperiment object containing marker intensities
and marker positivity}

\item{markers}{Character vector of markers to plot and the order in which
they should be plotted}

\item{celltypes}{Character vector of cell types to plot and the order in
which they should be plotted}

\item{parenttypes}{Character vector of parent cell types to plot and the
order in which they should be plotted}

\item{celltype_colname}{Column name in colData containing cell type
information (default: "HierarchyLevel4")}

\item{parent_colname}{Column name in colData containing parent cell type
information (default: "HierarchyLevel2")}

\item{value}{The value to plot. Either "expression" (default) or
"proportion"}
}
\value{
A ggplot2 object
}
\description{
Plot a heatmap of mean marker intensities or proportions
}
