% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarise_segmentation_stats.R
\name{summarise_segmentation_stats}
\alias{summarise_segmentation_stats}
\title{Summarise segmentation statistics from GeoJSON file}
\usage{
summarise_segmentation_stats(
  geojson_file,
  compartments = c("Cell", "Nucleus"),
  compartment_measurements = c("Area", "Length", "Circularity", "Solidity",
    "Max diameter", "Min diameter"),
  other_measurements = c("Nucleus/Cell area ratio"),
  summary_funcs = list(Mean = mean, Median = median)
)
}
\arguments{
\item{geojson_file}{Path to the GeoJSON file containing segmentation data}

\item{compartments}{Character vector of compartments to summarise (default:
c("Cell", "Nucleus"))}

\item{compartment_measurements}{Character vector of measurements to summarise
for each compartment (default: c("Area", "Length", "Circularity", "Solidity",
"Max diameter", "Min diameter"))}

\item{other_measurements}{Character vector of additional measurements to
summarise (default: c("Nucleus/Cell area ratio"))}

\item{summary_funcs}{List of functions to summarise the measurements
(default: list(Mean = mean, Median = median))}
}
\value{
A data.frame object
}
\description{
Take a GeoJSON file containing cell segmentation output and
summarise statistics, returning an output table. If measurements are not
available, the function will return only the cell counts.
}
